Solve day 01 part 1
God I am not looking forward to the next part however. It seems I've
really kludged myself into a deep, dark hole from which I will find it
very difficult to get myself out of.

I'm also not at the level where I can download the inputs automatically
so I'm staying put with manually copying them in to the inputs/
directory. What a fool. Rust feels so hard. I'm going to regret this
very quickly. God how I wish I was doing this in Python or something
that I'd feel more comfortable in. But, rules are rules. Rust for life
is the go for the next 25 days. Hoo boy.